More precise types
Instead of just saying SomeException, say what we really mean.  There
are a class of GHC exceptions we will call PseudoException that this
library is not interested in catching (since we could never know that an
action does not produce them anyway, they are not equivalent to return
values like exceptions thrown by throwIO are).

Get rid of the syncIO helper since errors doesn't need it any more and
it had a less-precise type anyway.