Simplify fromIO' and allow it to catch PseudoException
If the type specified is a PseudoException for fromIO' previously it
would compile but not catch anything.  While this is an odd edge
case, if you ask to catch a type, the library should catch it for you.

This rewrite has the added benefit of making fromIO' much simpler to
read.